[SPARK-47172] Moving ByteBufferWriteableChannel to its own class
